@@ -24,7 +24,6 @@ import java.net.URL;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.Iterator;
 import java.util.List;
 
 import org.codehaus.plexus.classworlds.AbstractClassWorldsTestCase;
@@ -65,7 +64,6 @@ public class ClassRealmImplTest
 
     @Test
     public void testLocateSourceRealm_NoImports()
-        throws Exception
     {
         ClassRealm realm = new ClassRealm( this.world, "foo", null );
 
@@ -445,9 +443,9 @@ public class ClassRealmImplTest
     }
 
     @Test
-    public void testLoadClass_Java11()
+    public void testLoadClassWithModuleName_Java9()
     {
-        final ExtendedClassRealm mainRealm = new ExtendedClassRealm(world);
+        final ExtendedClassRealm mainRealm = new ExtendedClassRealm( world );
         mainRealm.addURL( getJarUrl( "a.jar" ) );
         assertNotNull(mainRealm.simulateLoadClassFromModule( "a.A" ));
     }
@@ -472,14 +470,14 @@ public class ClassRealmImplTest
 
         assertEquals( baseUrl, subUrl );
 
-        List<String> urls = new ArrayList<String>();
-        for ( Iterator<URL> it = Collections.list( sub.getResources( resource ) ).iterator(); it.hasNext(); )
+        List<String> urls = new ArrayList<>();
+        for ( URL url : Collections.list( sub.getResources( resource ) ) )
         {
-            String path = it.next().toString();
+            String path = url.toString();
             path = path.substring( path.lastIndexOf( '/', path.lastIndexOf( ".jar!" ) ) );
             urls.add( path );
         }
-        assertEquals( Arrays.asList( new String[] { "/a.jar!/common.properties", "/b.jar!/common.properties" } ), urls );
+        assertEquals( Arrays.asList( "/a.jar!/common.properties", "/b.jar!/common.properties" ), urls );
     }
 
     @Test
@@ -502,16 +500,20 @@ public class ClassRealmImplTest
 
         List<URL> urls = Collections.list( child.getResources( resource ) );
         assertNotNull( urls );
-        assertEquals( Arrays.asList( new URL[] { childUrl, parentUrl } ), urls );
+        assertEquals( Arrays.asList( childUrl, parentUrl ), urls );
     }
 
-    // simulate new loadClass(Module,String) from java11
-    // it is reversed in terms of inheritance but enables to simulate the same behavior in these tests
-    private class ExtendedClassRealm extends ClassRealm
+    /**
+     * Simulates new {@code java.lang.ClassLoader#findClass(String,String)} introduced with Java 9.
+     * It is reversed in terms of inheritance but enables to simulate the same behavior in these tests.
+     * @see <a href="https://docs.oracle.com/en/java/javase/11/docs/api/java.base/java/lang/ClassLoader.html#findClass(java.lang.String,java.lang.String)">ClassLoader#findClass(String,String)</a>
+     */
+    private static class ExtendedClassRealm extends ClassRealm
     {
+        
         public ExtendedClassRealm(final ClassWorld world)
         {
-            super( world, "java11", Thread.currentThread().getContextClassLoader() );
+            super( world, "java9", Thread.currentThread().getContextClassLoader() );
         }
 
         public Class<?> simulateLoadClassFromModule(final String name)
